Move truncation logic to tools
